Ignore the rename text
It looks as if the field is set in case the left-hand item is already named, even if the current operation is not a rename.

We furthermore assume that skipping this check will not lead to unintended behavior because the Minecraft anvil appears to treat renames and enchantments as mutually exclusive.